from-scratch-transformer-nmt/requirements.txt
numpy
matplotlib
torch
transformer
dataset

